Making gRPC interface more specific
