--- mizar/proto/mizar/proto/nets.proto
@@ -1,32 +1,0 @@
-syntax = "proto3";
-
-package mizar;
-
-import "google/protobuf/empty.proto";
-
-service NetsService {
-    rpc CreateNet(NetMessage) returns (google.protobuf.Empty) {}
-    rpc UpdateNet(NetMessage) returns (google.protobuf.Empty) {}
-    rpc ReadNet(NetId) returns (NetsResponse) {}
-    rpc DeleteNet(NetId) returns (google.protobuf.Empty) {}
-    rpc ResumeNet(NetId) returns (google.protobuf.Empty) {}
-}
-
-message NetMessage {
-  string Name = 1;
-  string Ip = 2;
-  string Prefix = 3;
-  string Vni = 4;
-  string Vpc = 5;
-  string Status = 6;
-  string Bouncers = 7;
-}
-
-message NetId {
-  string Id = 1;
-}
-
-message NetsResponse {
-    string json_returncode = 1;
-}
-
