--- mizar/proto/mizar/proto/vpcs.proto
@@ -1,29 +1,0 @@
-syntax = "proto3";
-package mizar;
-
-import "google/protobuf/empty.proto";
-
-service VpcsService {
-    rpc CreateVpc(VpcMessage) returns (google.protobuf.Empty) {}
-    rpc UpdateVpc(VpcMessage) returns (google.protobuf.Empty) {}
-    rpc ReadVpc(VpcId) returns (VpcsResponse) {}
-    rpc DeleteVpc(VpcId) returns (google.protobuf.Empty) {}
-    rpc ResumeVpc(VpcId) returns (google.protobuf.Empty) {}
-}
-
-message VpcMessage {
-  string Name = 1;
-  string Ip = 2;
-  string Prefix = 3;
-  string Vni = 4;
-  string Dividers = 5;
-  string Status = 6;
-}
-
-message VpcId {
-  string Id = 1;
-}
-
-message VpcsResponse {
-    string json_returncode = 1;
-}
